pyproject.toml
[tool.poetry]
name = "crm-"
version = "0.1.0"
description = ""
authors = ["aigerim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
python-decouple = "^3.8"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
